+++ examples/hex_grid.rs
@@ -91,7 +91,7 @@ fn setup_grid(
                         text: Text::from_section(
                             format!("{},{}", hex.x, hex.y),
                             TextStyle {
-                                font_size: 8.0,
+                                font_size: 7.0,
                                 color: Color::BLACK,
                                 ..default()
                             },
@@ -203,10 +203,9 @@ fn hexagonal_plane(hex_layout: &HexLayout) -> Mesh {
         .facing(Vec3::Z)
         .with_scale(Vec3::splat(0.95))
         .build();
-    let mut mesh = Mesh::new(PrimitiveTopology::TriangleList);
-    mesh.insert_attribute(Mesh::ATTRIBUTE_POSITION, mesh_info.vertices);
-    mesh.insert_attribute(Mesh::ATTRIBUTE_NORMAL, mesh_info.normals);
-    mesh.insert_attribute(Mesh::ATTRIBUTE_UV_0, mesh_info.uvs);
-    mesh.set_indices(Some(Indices::U16(mesh_info.indices)));
-    mesh
+    Mesh::new(PrimitiveTopology::TriangleList)
+        .with_inserted_attribute(Mesh::ATTRIBUTE_POSITION, mesh_info.vertices)
+        .with_inserted_attribute(Mesh::ATTRIBUTE_NORMAL, mesh_info.normals)
+        .with_inserted_attribute(Mesh::ATTRIBUTE_UV_0, mesh_info.uvs)
+        .with_indices(Some(Indices::U16(mesh_info.indices)))
 }