+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added `ColumnMeshBuilder::with_scale` option to scale generated mesh (#121)
 * Added `ColumnMeshBuilder::with_rotation` option to rotate generated mesh (#121)
 * Mesh transformation follow the SRT order of operations (#121)
+* Bumped `bevy` deps to 0.12.x (#123)
 
 ## 0.11.0